requirements.txt
fastapi[standard]~=0.115.8
uvicorn~=0.34.0
line-bot-sdk~=3.16.0
aift~=1.3.2
pydantic_settings~=2.9.1